+++ pyrai/dispatcher/api.py
@@ -6,119 +6,8 @@ import IPython
 import plotly.graph_objects as go
 from dateutil.parser import isoparse
 from pytimeparse.timeparse import timeparse
-
-class FleetParams(object):
-    """
-    Class used to set fleet parameters for simulations.
-
-    Attributes:
-        max_wait (str): The max wait time.
-        max_delay (str): The max delay time.
-        unlocked_window (str): The unlocked window time.
-        close_pickup_window (str): The close pickup window time.
-    """
-
-    def __init__(self, max_wait, max_delay, unlocked_window, close_pickup_window):
-        """
-        Initializes a new FleetParams Object.
-
-        Args:
-            max_wait (str): The max wait time.
-            max_delay (str): The max delay time.
-            unlocked_window (str): The unlocked window time.
-            close_pickup_window (str): The close pickup window time.
-        """        
-
-        self.max_wait = max_wait
-        self.max_delay = max_delay
-        self.unlocked_window = unlocked_window
-        self.close_pickup_window = close_pickup_window
-
-    def todict(self):
-        """
-        Converts the FleetParams object to a python dictionary.
-        
-        Returns:
-            dict: A dictionary representation of self.
-        """
-
-        return {
-            'max_wait': self.max_wait,
-            'max_delay': self.max_delay,
-            'unlocked_window': self.unlocked_window,
-            'close_pickup_window': self.close_pickup_window
-        }
-
-class Defaults:
-    """
-    Class used to set default values for various parameters
-    used by the API.
-    """
-
-    BASE_URL = "https://api.routable.ai"
-    VISUALIZATION_URL = "https://dashboard.routable.ai/pyraimap?start={start}&end={end}&api_key={api_key}&fleet_key={fleet_key}"
-    DEFAULT_CAPACITY = 6
-    DEFAULT_DIRECTION = 0
-    DEFAULT_PARAMS = FleetParams(
-            max_wait="3m",
-            max_delay="6m",
-            unlocked_window="2m",
-            close_pickup_window="1s"
-        )
-
-
-class StatusError(Exception):
-    """
-    Error raised for responses that are not 200s.
-
-    Attributes:
-        resp (dict): The response.
-        status (int): The status of the response, should be nonzero
-            if error is raised.
-        error (string): Description of the error.
-    """
-    def __init__(self, resp=None, status=None, error=None):
-        """
-        Initializes a StatusError.
-
-        Args:
-            resp (dict, optional): The response. Defaults to None.
-            status (int, optional): The status of the response, should
-                be nonzero if error is raised. Defaults to None.
-            error (string, optional): Description of the error. Defaults to None.
-        """
-
-        if resp is not None:
-            self.status = resp.get('status')
-            self.error = resp.get('error')
-        else:
-            self.status = status
-            self.error = error
-
-    def __str__(self):
-        return self.error
-
-
-class Endpoints:
-    """
-    Class used to specify API endpoints.
-    """
-
-    CREATE_SIM_FLEET = "/dispatcher/simulation/create"
-    CREATE_LIVE_FLEET = "/dispatcher/live/create"
-    MAKE_VEHICLE_ONLINE = "/dispatcher/vehicle/online"
-    MAKE_VEHICLE_OFFLINE = "/dispatcher/vehicle/offline"
-    UPDATE_VEHICLE = "/dispatcher/vehicle/update"
-    REMOVE_VEHICLE = "/dispatcher/vehicle/remove"
-    GET_VEHICLE_INFO = "/dispatcher/vehicle"
-    ADD_REQUEST = "/dispatcher/request/add"
-    GET_REQUEST = "/dispatcher/request"
-    CANCEL_REQUEST = "/dispatcher/request/cancel"
-    COMPUTE_ASSIGNMENTS = "/dispatcher/assignments"
-    SET_PARAMS = "/dispatcher/params"
-    FORWARD_SIMULATE = "/dispatcher/simulation/forward"
-    GRAPHQL = "/graphql"
-
+from .structures import *
+from pyrai.helpers import to_rfc3339
 
 class Metrics:
     """
@@ -432,123 +321,12 @@ class Fleet(object):
             return StatusResponse(resp=resp)
         else:
             raise StatusError(resp=resp)
+
+    from .endpoints import make_vehicle_online
+        
+    from .endpoints import make_vehicle_offline
     
-    def make_vehicle_online(self, vid, location, capacity):
-        """
-        Attempts to make vehicle online.
-
-        Args:
-            vid (int): The vehicle ID.
-            location (Location): The vehicle location.
-            capacity (int): The vehicle capacity.
-
-        Raises:
-            StatusError: If unsuccessful.
-
-        Returns:
-            StatusReponse: If successful.
-        """
-
-        if datetime.datetime.now() > self.end_time:
-            self.end_time = datetime.datetime.now()
-
-        url = self.build_url(Endpoints.MAKE_VEHICLE_ONLINE)
-        payload = {
-            "location": location.todict(),
-            "id": vid,
-            'capacity': capacity,
-            'user_key': self.user_key.todict()
-        }
-        r = requests.post(url, data=json.dumps(payload))
-        resp = r.json()
-
-        if r.status_code == 200:
-            return StatusResponse(resp=resp)
-        else:
-            raise StatusError(resp=resp)
-        
-    def make_vehicle_offline(self, vid, location):
-        """
-        Attempts to take vehicle offline.
-
-        Args:
-            vid (int): The vehicle ID.
-            location (Location): The vehicle location.
-
-        Raises:
-            StatusError: If unsuccessful.
-
-        Returns:
-            StatusReponse: If successful.
-        """
-        if datetime.datetime.now() > self.end_time:
-            self.end_time = datetime.datetime.now()
-
-        url = self.build_url(Endpoints.MAKE_VEHICLE_OFFLINE)
-        payload = {
-            'location': location.todict(),
-            'id': vid,
-            'user_key': self.user_key.todict()
-        }
-        r = requests.post(url, data = json.dumps(payload))
-        resp = r.json()
-        
-        if r.status_code == 200:
-            return StatusResponse(resp=resp)
-        else:
-            raise StatusError(resp=resp)
-    
-    def update_vehicle(self, vid, location, event, direction=Defaults.DEFAULT_DIRECTION, event_time=None, req_id=None):
-        """
-        Attempts to update a vehicle.
-
-        Args:
-            vid (int): The unique vehicle ID
-            location (Location): The vehicle location.
-            direction (float): Angle in radians clockwise away from true north
-            event (VehicleEvent): Describes the current event for the vehicle. 
-                pickup occurs when the vehicle has picked up a request. 
-                dropoff occurs when the vehicle has dropped of a request. 
-                progress should be set when the vehicle is moving to service a request, 
-                either picking up or dropping off. The vehicle should be marked as 
-                unassigned when it is is not assigned to any requests.
-            event_time (datetime.datetime, optional): Time at which the vehicle update has occurred. 
-                Set to datetime.datetime.now() if not provided. Defaults to None.
-            req_id (int, optional): The unique ID of request the vehicle is servicing. 
-                If the vehicle is unassigned, this may be omitted. Defaults to None.
-
-        Returns:
-            Vehicle: If successful.
-
-        Raises:
-            StatusError: If unsucessful.
-        """
-        if event_time is None:
-            event_time = datetime.datetime.now()
-
-        if event_time > self.end_time:
-            self.end_time = event_time
-
-        url = self.build_url(Endpoints.UPDATE_VEHICLE)
-        payload = {
-            'id': vid,
-            'location': location.todict(),
-            'direction': direction,
-            'event_time': to_rfc3339(event_time),
-            'event': event,
-            'user_key': self.user_key.todict()
-        }
-        
-        if req_id is not None:
-            payload['req_id'] = req_id
-        
-        r = requests.post(url, data = json.dumps(payload))
-        resp = r.json()
-
-        if r.status_code == 200:
-            return Vehicle.fromdict(self, resp)
-        else:
-            raise StatusError(resp = resp)
+    from .endpoints import update_vehicle
 
     def remove_vehicle(self, vid, location):
         """
@@ -880,211 +658,6 @@ class Fleet(object):
                         mode='lines+markers',
                         name=metric))
         return figure
-        
-
-
-def to_rfc3339(dt):
-    """
-    Converts datetime objects to RFC3339 string for use with the API.
-
-    Args:
-        dt (datetime.datetime): A datetime.datetime object.
-
-    Returns:
-        string: An RFC3339 representation of dt.
-    """
-    return dt.astimezone(datetime.timezone.utc).isoformat()[:-6] + "Z"
-
-class Vehicle(object):
-    """ 
-    Class used to represent vehicles.
-
-    Attributes:
-        fleet (Fleet): the fleet the vehicle is a part of.
-        veh_id (int): the unique ID of the vehicle.
-        location (Locaiton): the location of the vehicle.
-        assigned (boolean): True if vehicle is assigned, false if not.
-        req_ids (list[int]): list of request IDs of assigned requests.
-        events (list[Event]): list of events this vehicle is assigned to.
-
-    """
-    def __init__(self, fleet, veh_id, location, assigned, req_ids, events):
-        """
-        Initializes a vehicle object
-
-        Args:
-            fleet (Fleet): the fleet the vehicle is a part of.
-            veh_id (int): the unique ID of the vehicle.
-            location (Locaiton): the location of the vehicle.
-            assigned (boolean): True if vehicle is assigned, false if not.
-            req_ids (list[int]): list of request IDs of assigned requests.
-            events (list[Event]): list of events this vehicle is assigned to.
-        """
-
-        self.fleet = fleet
-        self.veh_id = veh_id
-        self.location = location
-        self.assigned = assigned
-        self.req_ids = req_ids
-        self.events = events
-    
-    @staticmethod
-    def fromdict(fleet, d):
-        """
-        Converts a python dictionary into a Vehicle object.
-
-        Args:
-            fleet (Fleet): The fleet the vehicle is part of.
-            d (dict): The dictionary with the vehicle parameters.
-
-        Returns:
-            Vehicle: A vehicle object with the parameters the dictionary specifes.
-        """
-
-        return Vehicle(
-            fleet,
-            d.get('veh_id'),
-            Location.fromdict(d.get('location')),
-            d.get('assigned'),
-            d.get('req_ids'),
-            [Event.fromdict(e) for e in d.get('events')]
-        )
-
-    def todict(self):
-        """
-        Converts the Vehicle object to a python dictionary.
-
-        Returns:
-            dict: A dictionary representation of self.
-        """
-
-        return {
-            'fleet': self.fleet.todict(),
-            'veh_id': self.veh_id,
-            'location': self.location.todict(),
-            'assigned': self.assigned,
-            'req_ids': self.req_ids,
-            'events': [e.todict() for e in self.events]
-        }
-
-    def __str__(self):
-        return str(self.todict())
-
-    def make_online(self, location=None, capacity=Defaults.DEFAULT_CAPACITY):
-        """
-        Makes vehicle online.
-
-        Args:
-            location (Lovation, optional): Location of vehicle, uses self.location
-                if nothing is provided. Defaults to None.
-            capacity (int, optional): The capacity of the vehicle.
-                Defaults to Defaults.DEFAULT_CAPACITY.
-
-        Returns:
-            StatusResponse: If successful.
-        
-        Raises:
-            StatusError: If unsuccessful.
-        """
-
-        if location is None:
-            location = self.location
-
-        return self.fleet.make_vehicle_online(self.veh_id, location, capacity)
-
-    def make_offline(self, location=None):
-        """
-        Makes vehicle offline.
-
-        Args:
-            location (Lovation, optional): Location of vehicle, uses self.location
-                if nothing is provided. Defaults to None.
-
-        Returns:
-            StatusResponse: If successful.
-        
-        Raises:
-            StatusError: If unsuccessful.
-        """
-
-        if location is None:
-            location = self.location
-
-        return self.fleet.make_vehicle_offline(self.veh_id, location)
-
-    def update(self, 
-        event,
-        req_id=None,
-        location=None, 
-        direction=Defaults.DEFAULT_DIRECTION, 
-        event_time=None):
-        """
-        Updates the vehicle. Note that this mutates the vehicle, so nothing is returned.
-
-        Args:
-            location (Location, optional): The vehicle location, set to self.location
-                if nothing is provided. Defaults to None.
-            direction (float, optional): Angle in radians clockwise away from true north.
-                Defaults to Defaults.DEFAULT_DIRECTION.
-            event (VehicleEvent): Describes the current event for the vehicle. 
-                pickup occurs when the vehicle has picked up a request. 
-                dropoff occurs when the vehicle has dropped of a request. 
-                progress should be set when the vehicle is moving to service a request, 
-                either picking up or dropping off. The vehicle should be marked as 
-                unassigned when it is is not assigned to any requests.
-            event_time (datetime.datetime, optional): Time at which the vehicle update has occurred. Set to datetime.datetime.now() if not provided. Defaults to None.
-            req_id (int, optional): The unique ID of request the vehicle is servicing. 
-                If the vehicle is unassigned, this may be omitted. Defaults to None.
-
-        Returns:
-            None
-
-        Raises:
-            StatusError: If unsuccessful.
-        """
-        
-        if location is None:
-            location = self.location
-
-        updated_veh = self.fleet.update_vehicle(
-            vid=self.veh_id,
-            location=location,
-            direction=direction,
-            event_time=event_time,
-            req_id=req_id,
-            event=event
-        )
-
-        self.location = updated_veh.location
-        self.assigned = updated_veh.assigned
-        self.req_ids = updated_veh.req_ids
-        self.events = updated_veh.events
-
-        return
-
-    def remove(self, location=None):
-        """
-        Removes the vehicle.
-
-        Args:
-            location (Location, optional): The location of the vehicle,
-                will be set to self.location if nothing is provided.
-                Defaults to None.
-
-        Returns:
-            StatusResponse: If successful.
-
-        Raises:
-            StatusError: If unsuccessful.
-        """
-        
-        if location is None:
-            location = self.location
-
-        return self.fleet.remove_vehicle(self.veh_id, location)
-
-    def __repr__(self):
-        return str(self.todict())
 
 
 class UserKey(object):
@@ -1116,58 +689,6 @@ class UserKey(object):
         return {
             'api_key': self.api_key,
             'fleet_key': self.fleet_key
-        }
-
-    def __str__(self):
-        return str(self.todict())
-
-
-class Location(object):
-    """
-    Class for representing locations.
-
-    Attributes:
-        lat (float): The latitude.
-        lng (float): The longitude.
-    """
-
-    def __init__(self, lat, lng):
-        """
-        Initializes a location object
-
-        Args:
-            lat (float): The latitude.
-            lng (float): The longitude.
-        """
-
-        self.lat = lat
-        self.lng = lng
-    
-    @staticmethod
-    def fromdict(d):
-        """
-        Converts a dict object into a Location object.
-
-        Args:
-            d (dict): The dictionary to convert.
-
-        Returns:
-            Location: A Location object with attributes set by
-                fields in the dictionary.
-        """
-        return Location(d.get('lat'), d.get('lng'))
-
-    def todict(self):
-        """
-        Converts a Location object to a python dictionary.
-
-        Returns:
-            dict: A dictionary representation of self.
-        """
-
-        return {
-            'lat': self.lat,
-            'lng': self.lng
         }
 
     def __str__(self):
@@ -1295,66 +816,6 @@ class VehicleEvent:
     PROGRESS = "progress"
     UNASSIGNED = "unassigned"
 
-class Event(object):
-    """
-    Class used for representing events.
-
-    Attributes:
-        req_id (int): the ID of the request corresponding to this event
-        location (Location): the event Location
-        time (datetime.datetime): the event time.
-        event (VehicleEvent): the vehicle event corresponding to the event.
-    """
-    def __init__(self, req_id, location, time, event):
-        """
-        Initializes an Event object.
-
-        Args:
-            req_id (int): the ID of the request corresponding to this event
-            location (Location): the event Location
-            time (datetime.datetime): the event time.
-            event (VehicleEvent): the vehicle event corresponding to the event.
-        """
-        self.req_id = req_id
-        self.location = location
-        self.time = time
-        self.event = event
-
-    @staticmethod
-    def fromdict(d):
-        """
-        Converts a python dictionary to an Event object.
-
-        Args:
-            d (dict): The dictionary to convert.
-
-        Returns:
-            Event: An event with the parameters set by the values in d.
-        """
-        return Event(
-            d.get('req_id'),
-            Location.fromdict(d.get('location')),
-            isoparse(d.get('time')),
-            d.get('event')
-        )
-
-    def todict(self):
-        """
-        Converts an Event to a python dictionary.
-
-        Returns:
-            dict: A dictionary representation of self.
-        """
-        return {
-            'req_id': self.req_id,
-            'location': self.location.todict(),
-            'time': to_rfc3339(self.time),
-            'event': self.event
-        }
-    
-    def __str__(self):
-        return str(self.todict())
-
 class Notification(object):
     """
     Class for representing notifications.
@@ -1468,49 +929,6 @@ class NotificationData(object):
     def __str__(self):
         return str(self.todict())
 
-class StatusResponse(object):
-    """
-    Class used for representing Status Responses.
-
-    Attributes:
-        resp (dict): The response.
-        status (int): The status of the response.
-        error (string): Description of the error.
-
-    """
-    def __init__(self, resp=None, status=None, error=None):
-        """
-        Initializes a StatusResponse Object
-
-        Args:
-            resp (dict, optional): The response. Defaults to None.
-            status (int, optional): The status of the response. Defaults to None.
-            error (str, optional): Description of the error. Defaults to None.
-        """
-        if resp is not None:
-            self.status = resp.get('status')
-            self.error = resp.get('error')
-        else:
-            self.status = status
-            self.error = error
-
-    def todict(self):
-        """
-        Converts StatusResponse object to a python dictionary.
-
-        Returns:
-            dict: A dictionary representation of self.
-        """
-        return {
-            'status': self.status,
-            'error': self.error
-        }
-
-    def __str__(self):
-        return str(self.todict())
-
-    def __repr__(self):
-        return "Success!"
 
 class VehicleAssignments(object):
     """
