+++ pyrai/dispatcher/endpoints/fleet/__init__.py
@@ -1,0 +1,4 @@
+from .make_vehicle_online import make_vehicle_online
+from .make_vehicle_offline import  make_vehicle_offline
+from .update_vehicle import update_vehicle
+__all__ = ["make_vehicle_online", "make_vehicle_offline"]
