+++ pyrai/dispatcher/structures/__init__.py
@@ -1,0 +1,17 @@
+from .endpoints import Endpoints
+from .status_error import StatusError
+from .status_response import StatusResponse
+from .defaults import Defaults
+from .fleet_params import FleetParams
+from .event import Event
+from .location import Location
+from .vehicle import Vehicle
+__all__ = ["Endpoints", 
+"StatusResponse", 
+"StatusError", 
+"Defaults", 
+"FleetParams",
+"Event",
+"Location",
+"Vehicle"
+]
